Restore original API, move encrypt and decrypt into separate files
